Add Samples API.
Samples are obtained by adding types in a given order.
For each addition, the common type and sum type are updated.
The common type is the type in common to all samples.
The sum type is the aggregate sum of all samples.
One can get the last diff (w.r.t. the common type before the last addition), or all the diffs (w.r.t. the current common type).

There’s an example of how to use samples in Demo.re.